Twitter time line コンポーネント化
